// types.ts
export interface Bank {
  id: string;
  name: string;
  logo: React.FC<React.SVGProps<SVGSVGElement>>;
}

export type AccountStatus = 'Connected' | 'Sync Error' | 'Disconnected';

export interface Account {
  id: string;
  bankId: string;
  lastDigits: string;
  balance: number;
  status: AccountStatus;
  lastSync: string;
}